+++ pyclesperanto_prototype/_tier2/__init__.py
@@ -32,3 +32,9 @@ from ._sum_reduction_x import sum_reduction_x
 from ._top_hat_box import top_hat_box
 from ._top_hat_sphere import top_hat_sphere
 from ._touch_matrix_to_adjacency_matrix import touch_matrix_to_adjacency_matrix
+from ._z_position_of_maximum_z_projection import z_position_of_maximum_z_projection
+from ._z_position_of_maximum_z_projection import z_position_of_maximum_z_projection as arg_maximum_z_projection
+from ._z_position_of_minimum_z_projection import z_position_of_minimum_z_projection
+from ._z_position_of_minimum_z_projection import z_position_of_minimum_z_projection as arg_minimum_z_projection
+from ._z_position_projection import z_position_projection
+